Added Delete and Add Item Options
Added the options to delete and add items (items inside the groups) along with the dialogs needed. There is a bug when deleting the added items though, because they have no ID, as I don't have a way to generate it. The app will crash when trying to remove an added item.